Bug 1213146 - IonMonkey: MIPS: Modify the last two args of Assembler::bind to generic type. r=arai
---
 js/src/jit/mips-shared/Assembler-mips-shared.h | 2 +-
 js/src/jit/mips32/Assembler-mips32.cpp         | 2 +-
 js/src/jit/mips32/Assembler-mips32.h           | 2 +-
 js/src/jit/mips64/Assembler-mips64.cpp         | 2 +-
 js/src/jit/mips64/Assembler-mips64.h           | 2 +-
 5 files changed, 5 insertions(+), 5 deletions(-)

+++ js/src/jit/mips64/Assembler-mips64.cpp
@@ -256,7 +256,7 @@ Assembler::Bind(uint8_t* rawCode, AbsoluteLabel* label, const void* address)
 }
 
 void
-Assembler::bind(InstImm* inst, uint64_t branch, uint64_t target)
+Assembler::bind(InstImm* inst, uintptr_t branch, uintptr_t target)
 {
     int64_t offset = target - branch;
     InstImm inst_bgezal = InstImm(op_regimm, zero, rt_bgezal, BOffImm16(0));
